token implemented

+++ day44/backend/controllers/userController.js
@@ -1,3 +1,4 @@
+import { generateToken } from "../utilis/generate.js";
 import { createUserService, deleteUserService, getUserService, getUsersService, loginService, updateUserService } from "../services/userService.js"
 import { asyncHandler } from "../utilis/asyncHandler.js";
 
@@ -29,33 +30,47 @@ export const createUser=asyncHandler( async (req,res,next)=>{
 })
 
 export const loginUser=asyncHandler( async(req,res,next)=>{
-    const user=await loginService(req.body);
+    const user=await loginService(req.body,token);
     if(!user){
         const error=new Error("Invalid credentials");
         error.status=400;
         throw error; 
     }
     const {password,...safeUser}=user;
-    res.status(200).json({success:true,data:safeUser,msg:"user logging in"});
+    res.status(200).json({success:true,data:safeUser,token:user.token,msg:"user logging in"});
 })
 export const updateUser=asyncHandler( async(req,res,next)=>{
+    if(req.user.id !== Number(req.params.id)){
+        const error=new Error("Forbidden");
+        error.status=403;
+        throw error;
+    }
     const updated=await updateUserService(req.params.id,req.body);
+ 
     if(!updated){
         const error=new Error("user not found");
         error.status=404;
         throw error;
     }
+
     const {password,...safeUser}=updated;
     res.status(200).json({success:true,data:safeUser,msg:"user updated"});
 })
 
 export const deleteUser=asyncHandler( async (req,res,next)=>{
+    if(req.user.id !== Number(req.params.id)){
+        const error=new Error("Forbidden");
+        error.status=403;
+        throw error;
+    }
     const deleted=await deleteUserService(req.params.id);
+
     if(!deleted){
         const error=new Error("user not found");
         error.status=404;
         throw error; 
     }
+    
     const {password,...safeUser}=deleted;
     res.status(200).json({success:true,data:safeUser,msg:"user deleted"})
 })
